requirements.txt
discord
python-dotenv
quart
aiosqlite
aiohttp